+++ src/main/java/de/gematik/rbellogger/converter/RbelHtmlConverter.java
@@ -1,0 +1,88 @@
+/*
+ * Copyright (c) 2022 gematik GmbH
+ * 
+ * Licensed under the Apache License, Version 2.0 (the License);
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an 'AS IS' BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package de.gematik.rbellogger.converter;
+
+import de.gematik.rbellogger.data.RbelElement;
+import de.gematik.rbellogger.data.facet.RbelXmlFacet;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.tuple.Pair;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.*;
+import org.xml.sax.SAXException;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map.Entry;
+import java.util.Optional;
+
+@Slf4j
+class RbelHtmlConverter {
+    private static final String XML_TEXT_KEY = "text";
+
+    Optional<Document> parseHtml(String text) throws IOException, SAXException {
+        final String lowerCase = text.toLowerCase(Locale.ROOT);
+        if (!lowerCase.contains("<html") ||
+            !lowerCase.contains("</html>") ||
+            !lowerCase.startsWith("<")) {
+            return Optional.empty();
+        } else {
+            return Optional.of(Jsoup.parse(text));
+        }
+    }
+
+    void buildXmlElementForNode(Node branch, RbelElement parentElement, RbelConverter converter) {
+        final List<Entry<String, RbelElement>> childElements = new ArrayList<>();
+        parentElement.addFacet(RbelXmlFacet.builder()
+            .childElements(childElements)
+            .build());
+        for (Node childNode : branch.childNodes()) {
+            if (childNode instanceof TextNode) {
+                childElements
+                    .add(Pair.of(XML_TEXT_KEY,
+                        converter.convertElement(((TextNode) childNode).getWholeText(), parentElement)));
+            } else {
+                var rbelChildElement = new RbelElement(childNode.toString().getBytes(), parentElement);
+                childElements.add(Pair.of(childNode.nodeName(), rbelChildElement));
+                buildXmlElementForNode(childNode, rbelChildElement, converter);
+            }
+/*else if (child instanceof AbstractBranch) {
+                final RbelElement element = new RbelElement(
+                    ((AbstractBranch) child).asXML().getBytes(parentElement.getElementCharset()),
+                    parentElement);
+                buildXmlElementForNode((AbstractBranch) child, element, converter);
+                childElements.add(Pair.of(((AbstractBranch) child).getName(), element));
+            } else if (child instanceof Namespace) {
+                final String childXmlName = ((Namespace) child).getPrefix();
+                childElements
+                    .add(Pair.of(childXmlName, converter.convertElement(((Namespace) child).getText(), parentElement)));
+            } else if (child instanceof DefaultComment) {
+                // do nothing
+            } else {
+                throw new RbelException("Could not convert XML element of type " + child.getClass().getSimpleName());
+            }*/
+        }
+        if (branch instanceof Element) {
+            for (Attribute attribute : branch.attributes()) {
+                childElements.add(Pair.of(
+                    attribute.getKey(),
+                    converter.convertElement(attribute.getValue(), parentElement)));
+            }
+        }
+    }
+}