+++ src/test/java/de/gematik/rbellogger/converter/VauEpaConverterTest.java
@@ -21,7 +21,14 @@ import de.gematik.rbellogger.captures.PCapCapture;
 import de.gematik.rbellogger.configuration.RbelConfiguration;
 import de.gematik.rbellogger.converter.initializers.RbelKeyFolderInitializer;
 import de.gematik.rbellogger.data.RbelElement;
+import de.gematik.rbellogger.data.facet.RbelFacet;
+import de.gematik.rbellogger.data.facet.RbelVauEpaFacet;
+import de.gematik.rbellogger.key.RbelKey;
+import de.gematik.rbellogger.key.RbelVauKey;
+import de.gematik.rbellogger.renderer.RbelHtmlFacetRenderer;
 import de.gematik.rbellogger.renderer.RbelHtmlRenderer;
+import de.gematik.rbellogger.renderer.RbelHtmlRenderingToolkit;
+import j2html.tags.ContainerTag;
 import lombok.SneakyThrows;
 import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.BeforeEach;
@@ -30,14 +37,45 @@ import org.junit.jupiter.api.Test;
 import java.io.File;
 import java.io.IOException;
 import java.util.Base64;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Stream;
 
+import static de.gematik.rbellogger.renderer.RbelHtmlRenderingToolkit.*;
+import static j2html.TagCreator.div;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class VauEpaConverterTest {
 
     private PCapCapture pCapCapture;
     private RbelLogger rbelLogger;
+
+    {
+        RbelHtmlRenderer.registerFacetRenderer(
+            new RbelHtmlFacetRenderer() {
+                @Override
+                public boolean checkForRendering(RbelElement element) {
+                    return element.hasFacet(TestFacet.class);
+                }
+
+                @Override
+                public ContainerTag performRendering(RbelElement element, Optional<String> key, RbelHtmlRenderingToolkit renderingToolkit) {
+                    return ancestorTitle().with(
+                        vertParentTitle().with(
+                            childBoxNotifTitle(CLS_BODY).with(t2("Test Facet"))
+                                .with(div("Some Notes"))
+                        )
+                    );
+                }
+
+                @Override
+                public int order() {
+                    return 100;
+                }
+            }
+        );
+    }
 
     @BeforeEach
     public void setUp() {
@@ -47,6 +85,11 @@ public class VauEpaConverterTest {
             .build();
         rbelLogger = RbelLogger.build(new RbelConfiguration()
             .addInitializer(new RbelKeyFolderInitializer("src/test/resources"))
+            .addPostConversionListener((rbelElement, converter) -> {
+                if (rbelElement.hasFacet(RbelVauEpaFacet.class)) {
+                    rbelElement.addFacet(new TestFacet());
+                }
+            })
             .addCapturer(pCapCapture)
         );
         pCapCapture.close();
@@ -100,5 +143,25 @@ public class VauEpaConverterTest {
             .findRbelPathMembers("$.body.message.Envelope.Body.sayHello.arg0.text")
             .get(0).getRawStringContent())
             .isEqualTo("hello from integration client");
+    }
+
+    @Test
+    public void parentKeysForVauKeysShouldBeCorrect() {
+        assertThat(rbelLogger.getMessageHistory().get(7)
+            .findRbelPathMembers("$.body").get(0)
+            .getFacetOrFail(RbelVauEpaFacet.class)
+            .getKeyUsed())
+            .get()
+            .isInstanceOf(RbelVauKey.class)
+            .extracting(key -> ((RbelVauKey) key).getParentKey())
+            .extracting(RbelKey::getKeyName)
+            .isEqualTo("prk_vauClientKeyPair");
+    }
+
+    private class TestFacet implements RbelFacet {
+        @Override
+        public List<Map.Entry<String, RbelElement>> getChildElements() {
+            return List.of();
+        }
     }
 }